% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss_CS.R
\name{dissCS}
\alias{dissCS}
\title{Cauchy-Schwarz Divergence}
\usage{
dissCS(gmmobj1, gmmobj2)
}
\arguments{
\item{gmmobj1}{first \code{gmm}-like object.}

\item{gmmobj2}{second \code{gmm}-like object.}
}
\value{
computed distance value.
}
\description{
Cauchy-Schwarz divergence between two measures \eqn{P} and \eqn{Q} with corresponding 
probability density functions \eqn{p(x)} and \eqn{q(x)} is defined as
\deqn{
D_{CS}(P,Q) = -\log \left\brace \frac{\int p(x)q(x)dx}{\sqrt{ \int p(x)^2 dx \int q(x)^2 dx  }} \right\rbrace
}
which admits closed-form evaluation when \eqn{p(x)} (\code{gmmobj1}) and \eqn{q(x)} (\code{gmmobj2}) are two mixtures of Gaussians. 
Please note that input \code{gmm}-like objects are accepted only if these are outputs of 
functions from \pkg{gmmtools} package.
}
\examples{
\donttest{
# -------------------------------------------------------------
#                   Distance for Gaussian Mixtures
#
# Data 1 : use SMILEY data 'gensmiley()' function.
# Data 2 : SMILEY data is translated by +3 and rotated.
# Data 3 : SMILEY data is translated by +9 and rotated.
# -------------------------------------------------------------
## GENERATE DATA
#  set up
ndata = 10
ntot  = 3*ndata
rot   = qr.Q(qr(matrix(rnorm(4),ncol=2)))

#  generate
list_data = list()
for (i in 1:ndata){
  list_data[[i]]           = (T4cluster::gensmiley(n=150, sd=0.1)$data)
  list_data[[i+ndata]]     = (T4cluster::gensmiley(n=150, sd=0.1)$data)\%*\%rot + 3
  list_data[[i+(2*ndata)]] = (T4cluster::gensmiley(n=150, sd=0.1)$data)\%*\%rot + 9
}


## FIT GMM MODELS WITH K=4
list_gmm = list()
for (i in 1:ntot){
  list_gmm[[i]] = gmm(list_data[[i]], k=4)
}

## COMPUTE PAIRWISE DISTANCE
pdmat = array(0,c(ntot,ntot))
for (i in 1:(ntot-1)){
  gi = list_gmm[[i]]
  for (j in (i+1):ntot){
    gj = list_gmm[[j]]
    pdmat[i,j] <- pdmat[j,i] <- dissCS(gi, gj)
  }
}

## VISUALIZE
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
plot(list_data[[1]], main="SMILEY data", pch=19)
image(pdmat[,ntot:1], axes=FALSE, col=gray((0:64)/64),
      main="Cauchy-Schwarz Divergence")
par(opar)
}

}
\references{
\insertRef{kampa_closed-form_2011-2}{gmmtools}
}
\concept{diss}
