% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss_JR.R
\name{dissJR}
\alias{dissJR}
\title{Jensen-Renyi Divergence}
\usage{
dissJR(gmmobj1, gmmobj2)
}
\arguments{
\item{gmmobj1}{first \code{gmm}-like object.}

\item{gmmobj2}{second \code{gmm}-like object.}
}
\value{
computed distance value.
}
\description{
Jensen-Renyi Divergence of multiple measures \eqn{P_1,P_2,\ldots,P_N} is 
defined as
\deqn{
D_{JR}(P_1, P_2, \ldots, P_N) = H_\alpha (\sum \pi_i P_i) - \sum \pi_i H_\alpha (P_i)
}
where \eqn{\pi_i} are the weights of the measures, \eqn{H_\alpha (X)} is 
Renyi entropy of order \eqn{\alpha}. In our implementation, we use \eqn{\alpha=2} 
since the choice of order allows a closed-form expression and consider the case 
with two measures \eqn{P_1} and \eqn{P_2} only.
}
\examples{
\donttest{
# -------------------------------------------------------------
#                   Distance for Gaussian Mixtures
#
# Data 1 : use SMILEY data 'gensmiley()' function.
# Data 2 : SMILEY data is translated +3 and rotated.
# Data 3 : SMILEY data is translated +9 and rotated.
# -------------------------------------------------------------
## GENERATE DATA
#  set up
ndata = 10
ntot  = 3*ndata
gap   = 10
rot   = qr.Q(qr(matrix(rnorm(4),ncol=2)))

#  generate
ldata = list()
for (i in 1:ndata){
  ldata[[i]]           = (T4cluster::gensmiley(n=150, sd=0.1)$data)
  ldata[[i+ndata]]     = (T4cluster::gensmiley(n=150, sd=0.1)$data)\%*\%rot + 3
  ldata[[i+(2*ndata)]] = (T4cluster::gensmiley(n=150, sd=0.1)$data)\%*\%rot + 9
}


## FIT GMM MODELS WITH K=4
list_gmm = list()
for (i in 1:ntot){
  list_gmm[[i]] = gmm(ldata[[i]], k=4)
}

## COMPUTE PAIRWISE DISTANCE
distJR = array(0,c(ntot,ntot))
for (i in 1:(ntot-1)){
  gi = list_gmm[[i]]
  for (j in (i+1):ntot){
    gj = list_gmm[[j]]
    distJR[i,j] <- distJR[j,i] <- dissJR(gi, gj)
  }
}

## VISUALIZE
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
plot(ldata[[1]], main="SMILEY data", pch=19)
image(distJR[,ntot:1], axes=FALSE, col=gray((0:64)/64),
      main="Jensen-Renyi Divergence")
par(opar)
}

}
\references{
\insertRef{yang_closed-form_2009}{gmmtools}
}
\concept{diss}
