% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_density.R
\name{gmmdensity}
\alias{gmmdensity}
\title{Density Evaluation of Data given a Fitted GMM Model}
\usage{
gmmdensity(gmm, data)
}
\arguments{
\item{gmm}{an output of any GMM routines in our package of \code{gmmtools} class.}

\item{data}{an \eqn{(n\times p)} query points.}
}
\value{
a length-\eqn{n} vector of evaluated densities.
}
\description{
Given a GMM model of class \code{gmmtools} \eqn{P(x) = \sum_{k=1}^K w_k \mathcal{N}(x\vert \mu_k, \Sigma_k)} in \eqn{\mathbf{R}^p}, 
\code{gmmdensity} evaluates the density of given data \eqn{\lbrace x_1, x_2, \ldots, x_n \rbrace}.
}
\examples{
# -------------------------------------------------------------
#       Evaluate Density of GMM Fit from Smiley Data
# -------------------------------------------------------------
# STEP 1. Generate Data and Fit with 9 Components
mydata = T4cluster::genSMILEY()$data
myfit  = gmm(mydata, k=9)

# STEP 2. Generate a Grid and Evaluate Over Grid Points
grid.data = as.matrix(expand.grid(x=seq(from=-1.5,to=1.5,length.out=250),
                                  y=seq(from=-1.5,to=1.5,length.out=250)))
grid.prob = gmmdensity(myfit, grid.data)

\dontrun{
# STEP 3. Visualize with 'ggplot2' (* optional)
#    3-1. wrap as a single data.frame for use with ggplot2
grid.df = as.data.frame(cbind(grid.data, density=grid.prob))

#    3-2. plot it!
ggplot2::ggplot(grid.df,aes(x=x,y=y,z=density)) + 
    geom_raster(aes(fill=density)) + 
    geom_contour(colour="white") +
    scale_fill_viridis_c() +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    coord_fixed(xlim=c(-1.5,1.5), ylim=c(-1.5,1.5)) + 
    ggtitle("contour of 9-component GMM density")
}

}
\concept{util}
