% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_label.R
\name{gmmlabel}
\alias{gmmlabel}
\title{Label Prediction of Data given a Fitted GMM Model}
\usage{
gmmlabel(gmm, data)
}
\arguments{
\item{gmm}{an output of any GMM routines in our package of \code{gmmtools} class.}

\item{data}{an \eqn{(n\times p)} query points.}
}
\value{
a length-\eqn{n} vector of class labels.
}
\description{
Given new data with the fitted GMM models, predict the cluster labels for the newly provided data according to the fitted model. Please note that this function benefits from the coherent structures of the package itself so that the input \code{gmmobj} must be an output of any GMM-type routines in \pkg{gmmtools} package.
}
\examples{
# -------------------------------------------------------------
#            clustering with 'iris' dataset
# -------------------------------------------------------------
## PREPARE
data(iris)
X   = as.matrix(iris[,1:4])

## LEARN WITH TESTING DATA FOR K=3
cl3 = gmm(X, k=3)

## PREDICT LABEL FOR THE SAME DATA 
lab.test = gmmlabel(cl3, X)
lab.pred = cl3$cluster

## EMBEDDING WITH PCA
X2d    = Rdimtools::do.pca(X, ndim=2)$Y

## VISUALIZATION
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
plot(X2d, col=lab.pred, pch=19, main="original results")
plot(X2d, col=lab.test, pch=19, main="predicted label")
par(opar)

}
\concept{util}
