% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_medianL2.R
\name{medianL2}
\alias{medianL2}
\title{Median of GMMs under \eqn{L^2} Metric}
\usage{
medianL2(gmmlist, weight = NULL, ...)
}
\description{
Median of GMMs under \eqn{L^2} Metric
}
\examples{
# -------------------------------------------------------------
#                   Mixtures for SMILEY Data
#
# STEP 1. Generate 10 datasets with noise
# STEP 2. Fit GMM with varying number of components (k=3-8)
# STEP 3. Combine GMMs with uniform weights and Visualize
# -------------------------------------------------------------
# STEP 1. 10 datasets with noise
list_data = list()
for (i in 1:10){
  list_data[[i]] = T4cluster::genSMILEY(sd=0.25)
}

# STEP 2. Fit GMM with varying k
list_gmm = list()
for (i in 1:10){
  list_gmm[[i]] = gmmtools::gmm(list_data[[i]]$data, k=sample(3:8, 1))
}

# STEP 3. Find Median of GMMs under L2 metric
medgmm = medianL2(list_gmm)

# prepare grid and density evaluation
grid.data = expand.grid(x=seq(from=-1.5,to=1.5,length.out=250),
                        y=seq(from=-1.5,to=1.5,length.out=250))
grid.prob = gmmdensity(medgmm, grid.data)
grid.df   = as.data.frame(cbind(grid.data, density=grid.prob))

\dontrun{
# we use 'ggplot2' for visualization
require(ggplot2)
ggplot(grid.df, aes(x=x,y=y,z=density)) + 
  geom_raster(aes(fill=density)) + 
  geom_contour(colour="white") +
  scale_fill_viridis_c() +
  scale_x_continuous(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0)) +
  coord_fixed(xlim=c(-1.5,1.5), ylim=c(-1.5,1.5)) + 
  ggtitle("L2 Median of 10 GMMs")
}

}
\concept{combine}
