% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tcenter.R
\name{tcenter}
\alias{tcenter}
\title{t-Center of Total Square Loss}
\usage{
tcenter(gmmlist)
}
\arguments{
\item{gmmlist}{list of \eqn{N} \code{gmmtools} S3 objects of \code{gmm}-type.}
}
\value{
a \code{gmmtools} S3 object of \code{gmm}-type object.
}
\description{
Total Square Loss (TSL) is a class of total Bregman divergence generated by 
the square loss function and written as 
\deqn{
tSL(x,y) = \frac{(x-y)^2}{\sqrt{1 + 4y^2}}.
}
\code{tcenter} computes \eqn{t}-center of multiple GMMs with respect to 
the TSL, which gives a closed-form solution.
}
\examples{
# -------------------------------------------------------------
#                   Mixtures for SMILEY Data
#
# STEP 1. Generate 10 datasets with noise
# STEP 2. Fit GMM with varying number of components (k=3-9)
# STEP 3. Combine GMMs with the Method and Visualize
# -------------------------------------------------------------
# STEP 1. 10 datasets with noise
list_data = list()
for (i in 1:10){
  list_data[[i]] = T4cluster::gensmiley(sd=0.25)
}

# STEP 2. Fit GMM with varying k
list_gmm = list()
for (i in 1:10){
  list_gmm[[i]] = gmm(list_data[[i]]$data, k=sample(3:9, 1))
}

# STEP 3. Combine GMMs with the Methods and Visualize
# Compute t-center
tcenter = tcenter(list_gmm)

# prepare grid and density evaluation
npts = 250
grid.data = expand.grid(x=seq(from=-1.5,to=1.5,length.out=npts),
                        y=seq(from=-1.5,to=1.5,length.out=npts))
grid.prob = gmmdensity(grid.data, tcenter)
grid.df   = cbind(grid.data, density=grid.prob)

\dontrun{
# plot
ggplot2::ggplot(grid.df, aes(x=x,y=y,z=density)) + 
  geom_raster(aes(fill=density)) + 
  geom_contour(colour="white") +
  scale_fill_viridis_c() +
  scale_x_continuous(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0)) +
  coord_fixed(xlim=c(-1.5,1.5), ylim=c(-1.5,1.5)) + 
  ggtitle("t-center of 10 GMMs")
}  

}
\references{
\insertRef{liu_shape_2012}{gmmtools}
}
\concept{merge}
