% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrap1d.R
\name{wrap1d}
\alias{wrap1d}
\title{Wrap Univariate Gaussian Mixture Model}
\usage{
wrap1d(weight, mean, variance)
}
\arguments{
\item{weight}{a length-\eqn{K} vector of class weights that sum to 1.}

\item{mean}{a length-\eqn{K} vector of component means.}

\item{variance}{a length-\eqn{K} vector of component variances.}
}
\value{
a S3 object of \code{gmmtools} class.
}
\description{
Given a set of user-specified parameters, this function returns an object of 
S3 class \code{gmmtools} that represents a 1-dimensional Gaussian mixture model.
}
\examples{
\donttest{
# a simple example with 2 components
my_pi  = c(0.4, 0.6)
my_mu  = c(-1, 1)
my_var = c(2, 2)

# wrap with given parameters
gmm1d  = wrap1d(my_pi, my_mu, my_var)
}

}
\concept{util}
