% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrapmd.R
\name{wrapmd}
\alias{wrapmd}
\title{Wrap Multivariate Gaussian Mixture Model}
\usage{
wrapmd(weight, mean, variance)
}
\arguments{
\item{weight}{a length-\eqn{K} vector of class weights that sum to 1.}

\item{mean}{a \eqn{(K\times n)} matrix of component means.}

\item{variance}{a \eqn{(n\times n\times K)} array of component covariances.}
}
\value{
a S3 object of \code{gmmtools} class.
}
\description{
Given a set of user-specified parameters, this function returns an object of 
S3 class \code{gmmtools} that represents an \eqn{n}-dimensional Gaussian mixture model.
}
\examples{
\donttest{
# a simple example with 2 components in R^3
my_pi  = c(0.4, 0.6)
my_mu  = matrix(rnorm(2*3), nrow=2)
my_var = array(0,c(3,3,2))
for (i in 1:2){
  tmpmat = matrix(rnorm(10*3),ncol=3)
  my_var[,,i] = stats::cov(tmpmat)
}

# wrap with given parameters
gmm3d  = wrapmd(my_pi, my_mu, my_var)
}

}
\concept{util}
