% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_medianL2.R
\name{medianL2}
\alias{medianL2}
\title{Median of GMMs under \eqn{L_2} Metric}
\usage{
medianL2(gmmlist, weight = NULL, ...)
}
\arguments{
\item{gmmlist}{list of \eqn{N} \code{gmmtools} S3 objects of \code{gmm}-type.}

\item{weight}{weight for each mixture model. If \code{NULL} (default), uniform weights are set.}

\item{...}{extra parameters for iterative updates, including \describe{
\item{maxiter}{the number of iterations (default: 50).}
\item{abstol}{stopping criterion (default:1e-6).}
}}
}
\value{
a \code{gmmtools} S3 object of \code{gmm}-type object.
}
\description{
Given multiple GMMs \eqn{P_1, P_2, \ldots, P_N}, geometric median induced 
by \eqn{L_2} metric is defined as follows
\deqn{
\bar{P} = \underset{P}{\textrm{argmin}} \sum_{n=1}^N w_n d(P,P_n)
}
where \eqn{d(\cdot,\cdot)} is \eqn{L_2} distance and \eqn{w_n}'s are nonnegative 
numbers that sum to 1.
}
\examples{
# -------------------------------------------------------------
#                   Mixtures for SMILEY Data
#
# STEP 1. Generate 10 datasets with noise
# STEP 2. Fit GMM with varying number of components (k=3-9)
# STEP 3. Combine GMMs with uniform weights and Visualize
# -------------------------------------------------------------
# STEP 1. 10 datasets with noise
list_data = list()
for (i in 1:10){
  list_data[[i]] = T4cluster::gensmiley(sd=0.25)
}

# STEP 2. Fit GMM with varying k
list_gmm = list()
for (i in 1:10){
  list_gmm[[i]] = gmm(list_data[[i]]$data, k=sample(3:9, 1))
}

# STEP 3. Find Median of GMMs under L2 metric
medgmm = medianL2(list_gmm)

# prepare grid and density evaluation
grid.data = expand.grid(x=seq(from=-1.5,to=1.5,length.out=250),
                        y=seq(from=-1.5,to=1.5,length.out=250))
grid.prob = gmmdensity(grid.data, medgmm)
grid.df   = as.data.frame(cbind(grid.data, density=grid.prob))

\dontrun{
# we use 'ggplot2' for visualization
ggplot(grid.df, aes(x=x,y=y,z=density)) + 
  geom_raster(aes(fill=density)) + 
  geom_contour(colour="white") +
  scale_fill_viridis_c() +
  scale_x_continuous(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0)) +
  coord_fixed(xlim=c(-1.5,1.5), ylim=c(-1.5,1.5)) + 
  ggtitle("L2 Median of 10 GMMs")
}
 
}
\concept{merge}
